stable waypoint 2: add document pending, get a heads up pending
